+++ containers/Home/index.ts
@@ -1,0 +1,9 @@
+import { About } from './About/About'
+import { Contact } from './Contact/Contact'
+import { Language } from './Language/Language'
+import { Lib } from './Lib/Lib'
+import { Skill } from './Skill/Skill'
+import { Achieve } from './Achieve/Achieve'
+import { Service } from './Service/Service'
+
+export { About, Contact, Language, Lib, Skill, Service, Achieve }
